[Chore] #4 측정 화면을 ViewController에서 그리도록 변경
