+++ GyroData/MotionMode.swift
@@ -5,7 +5,7 @@
 //  Created by 천수현 on 2022/12/27.
 //
 
-enum MotionMode {
+enum MotionMode: CaseIterable {
     case accelerometer
     case gyroscope
 }
